requirements.txt
# Web Framework
Flask
streamlit

# AI/ML Libraries
langchain
torch
transformers

# Version Control
GitPython

# Configuration
python-dotenv